Started to make the bulkuploader more general
Not sure that this is the best approach
I have users first point to the R object that corresponds to the user data and rename this "originalcsv" (changing from miomap csv).
Then I have users specify the column names in their dataframe that correspond to the generic names we use in the function
I then use which statements in the function to point the function to the correct columns in the dataframe.
This works fine until makeCite...which does not work with these "which" functions to retrieve columns. Maybe because there is a difference in how Tibbles handle this from a general dataframe?